now that I have a working solution, refactor to organize into logical pieces
